Exercises 4.15-4.23 completed

+++ part4/bloglist/controllers/blogs.js
@@ -1,12 +1,14 @@
-const blogRouter = require('express').Router()
+const middleware = require('../utils/middleware')
+const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 
-blogRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+blogsRouter.get('/', async (request, response) => {
+  const blogs = await Blog
+    .find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
-blogRouter.get('/:id', async (request, response) => {
+blogsRouter.get('/:id', async (request, response) => {
   const blog = await Blog.findById(request.params.id)
   if (blog) {
     response.json(blog)
@@ -15,27 +17,46 @@ blogRouter.get('/:id', async (request, response) => {
   }
 })
 
-blogRouter.post('/', async (request, response) => {
-  const blog = new Blog(request.body)
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
+  const body = request.body
+  const user = request.user
 
-  if (blog.title === undefined || blog.url === undefined) {
+  if (body.title === undefined || body.url === undefined) {
     return response.status(400).end()
   }
 
-  if (blog.likes === undefined) {
-    blog.likes = 0
+  if (body.likes === undefined) {
+    body.likes = 0
   }
 
+  const blog = new Blog({
+    title: body.title,
+    author: body.title,
+    url: body.url,
+    likes: body.likes,
+    user: user.id
+  })
+
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
   response.status(201).json(savedBlog)
 })
 
-blogRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  const user = request.user
+
+  if (user.id !== blog.user.toString()) {
+    return response.status(401).json({ error: 'mismatched ids' })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
 
-blogRouter.put('/:id', async (request, response) => {
+blogsRouter.put('/:id', async (request, response) => {
   const blog = request.body
 
   const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
@@ -47,4 +68,4 @@ blogRouter.put('/:id', async (request, response) => {
   }
 })
 
-module.exports = blogRouter
+module.exports = blogsRouter
